server/requirements.txt
alembic==1.7.7
click==8.1.2
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
SQLAlchemy==1.4.35
Werkzeug==2.1.1
zipp==3.8.0
python-dotenv
Flask-MySQLdb==1.0.1
pymysql
flask-cors
flask-session
redis